[repository #165] refactor
