+++ cargo-smart-release/src/command/release/git.rs
@@ -6,15 +6,18 @@ use cargo_metadata::{
     camino::{Utf8Component, Utf8Path},
     Package,
 };
+
+use git_repository::easy::object;
 use git_repository::{
     hash::ObjectId,
-    lock, object,
+    lock,
     prelude::ReferenceAccessExt,
     refs::{self, file::loose::reference::peel},
 };
 
-use super::{Context, Options};
 use crate::command::release_impl::{tag_name_for, utils::will};
+
+use super::{Context, Options};
 
 fn is_top_level_package(manifest_path: &Utf8Path, shared: &git_repository::Easy) -> bool {
     manifest_path
